index.d.ts, src/index: Remove toXYZ(), fromMap and fromArray from Value
Values can be thought of as "output" types. Their sole purpose is for
them to be created in the mappings and passed up to the Rust host process
for storing in the database. Thus, we don't need to provide ways of working
any further with these values, e.g. by converting them to other types.