make keytar an optional dependency
it's sort of a shame to have all projects that make use of graph-cli
require libsecret when it's not necessary in the vast majority of cases.
this commit makes it an optional dependency and doesn't attempt to load
it until it's needed.

i don't anticipate this being a major problem - most of the time this
auth store is being used, it's in a desktop environment where libsecret
will already be present. conversely, most of the time libsecret is
absent it will be in a CI environment of some sort, where it's likely
not needed.