commands/init: Add --index-events flag, load ABIs properly
Event entities and fully built out handlers are now only generated when
`--index-events` is passed to `graph init`. Otherwise, an example entity
is generated, with mostly empty event handlers. One of the event
handlers includes instructive comments and a few examples of how to work
with entities and make contract calls.