compiler: Cache uploaded files using SHA1 instead of paths (#393)
Some projects generate data sources in the manifest from configuration
files and end up with _many_ identical ABIs and mappings in different
paths.

We used to cache uploaded files by `(path, ipfs-hash)` but that's not
good enough in this case. This changes it so we use `(sha1, ipfs-hash)`
as the cached information, which deduplicates uploads based on file
content rather than paths.