add support to load generated manifests
in order to use the same subgraph for multiple neworks (ex. mainnet, kovan, mainnet-ovm, etc.), it is currently
necessary to construct a fresh subgraph manifest for each network and supply them individually.

this PR is a slight modification to the CLI in order to resolve the manifest definition from a JS file. as long
as the javascript resolves its `module.exports` to the JSON of the manifest, custom code can be used to generate
the appropriate manifest

this should not break any existing compatibility with YAML or JSON loading.