fix: Remove ciclic -cli dependency
This is very important because before we would download
graph-cli from the latest version to run the tests against it.

However this is bad, because we could not be sure a new version
passes against the tests from the example folder, since the release
needed to be done before (it would go to master still pointing to
the old version).

This commit fixes the issue by pointing to the local graph-cli
so that the example tests fail before we do a release. This also
removes the necessity to bump the version everytime on these files.