fix(scaffold): correct entity ID format in generated tests (#2078)
* fix(scaffold): correct entity ID format in generated tests

When generating test files with --index-events, the entity ID format was
using the old string concatenation format ('-1') instead of the format
produced by concatI32() ('01000000'). This caused all generated tests to
fail out-of-the-box.

Changes:
- Update entity ID in test assertions from '-1' to '01000000' format
- Improve comment to explain concatI32() encoding with little-endian bytes
- Correct terminology from "default address" to "default transaction hash"

This fixes a bug introduced in commit 34aee498 (Oct 2022) when Bytes as ID
with @immutable was implemented. The mapping and schema templates were
updated but the test template was missed.

* add changeset

* update test

* add proper healthcheck to docker compose and wait

* lint

* pin postgres

---------

Co-authored-by: YaroShkvorets <shkvorets@gmail.com>